fix: reset the style of elements only when passing nullish param (#2798)
Previously, passing an empty array acts like passing a nullish
parameter. The style of all elements was reset.

Also update the JSDoc of some methods of `BpmnElementsRegistry` to
clarify what exactly happen when passing nullish or empty array as
parameter for the ids of the BPMN elements. Create the corresponding
integration tests to confirm the actual behavior.