refactor: remove some coupling between the "CSS" and "Style" API (#2825)
Previously, when applying the "CSS" API, the implementation ensured that
the cell's original style was cached (for any subsequent reset) without
any additional CSS classes. In fact, this was not necessary. Storing
styles in the cache with additional CSS classes information posed no
problem. When the style is reset, the value retrieved from the cache is
updated with the CSS classes currently applied to the cell. This works
just as well if there are no CSS classes as if there are. This
additional check has therefore been removed.

The tests verifying interactions between the "CSS" and "Style" APIs have
been improved. They confirm that the removal of this coupling is
legitimate:
  - Introduce a dedicated test file for these tests.
- Previously, they were stored in the file testing the style update at
template level.
- This change provides an explicit way of identifying API interaction
tests and a better separation of concerns.
- Add a dedicated test to check the impact of resetting the style after
a CSS and style update. This proves the uselessness of the coupling
described above.

covers #2803
closes #2743

---------

Co-authored-by: Souchet Céline <4921914+csouchet@users.noreply.github.com>