feat: add a method to only get BPMN semantic of elements by kind (#2841)
Complement the existing method that also retrieves DOM information. This
simplifies the API usage and reduce the amount of retrieved data when
the caller only needs to get the information from the model.

In addition:
- Simplify the existing `getElementsByKinds` implementation: use a flat
array instead of an array of array. This is then also consistent with
the implementation of the new method.
  - Use the new API in demo to simplify the code.
  - Rework the existing JSDoc for consistency