ci(release): eliminate bypassing of branch protection
The workflow commits to the main branch and the branch protection configured in the GitHub repository prevents this for all users.

Previously, the protection was temporarily bypassed and the user performing the commit required administrator permissions (due to the configuration of the branch protection).

Now protection is done using GitHub branch ruleset, and there's no need to bypass protection. In this way, workflow can be simplified.