docs: clarify the Pull Request template
The new template is aligned with the content of the recently updated contribution guide.
The aim is to ensure that we do not receive Pull Requests on non-priority topics and that the contribution respects our
guidelines.

In particular, the template now includes a checklist that allows contributors to validate that they are following the
contribution process.