fix(API): do not reset/remove when passing an empty string
This applies to the "remove all CSS classes" and "reset style" APIs.

These two methods are intended to apply to all elements when a nullish parameter is passed.
However, they previously applied to the empty string because the condition was checking the "falsy" state instead.

New tests have been added to reproduce the problem and validate the fix.