feat: introduce a dedicated function to get the version (#2965)
This simplifies the process of obtaining the version of
`bpmn-visualization`.

Previously, it was necessary to instantiate `BpmnVisualization` to
obtain the version, whereas this information was available without any
other prerequisites.
The version can now be retrieved directly by calling the new
`getVersion` function. The old `BpmnVisualization.getVersion` method has
therefore been marked as deprecated, and will be removed in the future.
In the API documentation, the new function and the existing `Version`
type are not categorized. The `Version` type was previously categorized
as "Initialization & Configuration", but in fact is not related to this
topic.

In addition, the TypeScript configuration dedicated to Typedoc has been
updated to take into account only the `src` folder and no longer the
demo folder, which is not part of the API documentation. This speeds up
documentation generation.