fix: ignore invalid event definition reference (#3018)
Previously, the parser assumed that the identifier reference in the
event definition was valid, i.e. that the identifier was defined at the
top level.
If this was not the case, the corresponding event definition was not
retrieved. The following code attempted to access properties with
undefined values.
This generated an error.
The parser now ignores the invalid reference and set the event definition
to the default value (none).