refactor: inject IconPainter when possible
Previously, the shapes relied on `IconPainterProvider` to get the `IconPainter` instance. This instance is now injected
at shape instantiation by `BpmnCellRenderer`. This reduces the adherence to a global state.

This is a first step to remove the `IconPainterProvider` global state. This may be done later as it would imply to:
  - introduce a new configuration options, for example a `render.iconPainter` property
  - change how the `BpmnGraph` is initialized and configured
  - manage the breaking changes introduced by the provider removal