fix: ensure numeric attributes are parsed as number (#3040)
The XML parser now ensures that the following elements have numeric
attributes
  - bounds of shapes
  - label bounds of shapes and edges
  - waypoints of edges

Previously, when the attribute contained too many decimals, the value
was converted to a string instead of a number. Existing tests exhibited
this behavior; they have been modified and enriched to cover all use
cases and show that the issue is now fixed.
A high-level parsing test has also been added to use the diagram that
was provided by the community to show the past problem.