refactor: inject `IconPainter` in Shapes (#3042)
Previously, the shapes relied on `IconPainterProvider` to get the
`IconPainter` instance. This instance is now injected at shape
instantiation by `BpmnCellRenderer`. This reduces the adherence to a
global state.

This is a first step to remove the `IconPainterProvider` global state.
This may be done later as it would imply to:
- introduce a new configuration options, for example a
`render.iconPainter` property
  - change how the `BpmnGraph` is initialized and configured
  - manage the breaking changes introduced by the provider removal

### Minor functional change

Previously, when a shape was rendered, the current instance of the
`IconPainter` was injected. If, between two renders, a new instance of
`IconPainter` was configured via `IconPainterProvider`, it was used in
the next render. This was a side-effect, not a desired feature, and was
not documented.
The interest is limited: we haven't yet seen a use case that requires
it. The existing example of modifying the user task icon doesn't need
it: in this case, we still want to display the customized icon.

With the changes proposed here, the behavior is modified. The
`IconPainter` instance is defined when the library is initialized, so
its modification in `IconPainterProvider` only applies to new instances
of the library. All rendering of a shape will always use the
`IconPainter` instance defined when the library was initialized.