refactor: directly override `shape.createSvgCanvas`
Previously, this was done by modifying the shape prototype.
The new implementation limits side effects on application that use mxGraph directly, and don't modify code we don't own.

The change is validated by integration tests that check that the bpmn-id and the CSS classes in the DOM are correctly
set on both shapes/edges and labels.
This is also validated with the elements-identification.html demo. The CSS class is correctly applied to the label.