refactor: directly override `svgCanvas2D.getTextCss`
Previously, this was done by modifying the svgCanvas2D prototype.
The new implementation limits side effects on applications that use mxGraph directly, and don't modify code the lib
doesn't own.

The mxGraph implementation overriding is required to ensure that the label pointer-events is set to 'none'.
New integration tests have been added to validate the old behavior and ensure that the refactoring has preserved it.
In particular, more tests check the labels of a greater number of BPMN element types, and the value of the "pointer-events"
property of labels is explicitly checked.

The overriding implementation has been simplified to reduce the copy/paste of mxGraph code. Instead, it now calls the
standard mxGraph implementation with specific settings.