FIX side effects: configure paintLoop to not update the iconConfig shared by all markers
This had a side effects on subsequent markers that needed to be filled.
Directly set the fillColor of the underlying mxGraph canvas, like this done in
  - paintParallelMultiInstanceIcon
  - paintSequentialMultiInstanceIcon

As for the multi instance icon, there is no need to make the fill configurable (via iconConfig in PaintParameter) as the
arrow of the loop must always be filled.