refactor!: remove unused BpmnCanvas.setIconOriginForIconOnBottomLeft
In addition, remove StyleDefault.SHAPE_ACTIVITY_FROM_CENTER_MARGIN that was only used in this method.
The StyleDefault value was only used in the removed method.

BREAKING CHANGES: remove unused elements
  - BpmnCanvas.setIconOriginForIconOnBottomLeft
  - StyleDefault.SHAPE_ACTIVITY_FROM_CENTER_MARGIN
There should be no impact for users of bpmn-visualization as there is no reason to use this former method.
No existing repository of the Process Analytics project was using it.