fix: remove extra spacing between 2 activity markers (#3052)
When there are 2 markers, the center was translated/shifted too much. We
assume that the width value of markers is
`SHAPE_ACTIVITY_MARKER_ICON_SIZE`. We want to have a spacing value
between the 2 markers of `SHAPE_ACTIVITY_MARKER_ICON_MARGIN`.

To have 2 markers horizontally centered and closed to each other, we
need to translate the first marker to the left and the second to the
right by half the value of the marker width. To add a space between the
2 markers, we need to translate both markers by half the value of the
global "spacing value". Previously, the translation value included the
full marker width which is too much.

Also include refactoring:
- Use more accurate name for the parameters of
`getMarkerIconOriginFunction`.