fix!: generate correct CSS class name for adhoc subprocess (#3058)
The CSS class name associated to the adhoc subprocess was incorrect. It
was `bpmn-sub-process-ad_hoc` instead of `bpmn-sub-process-adhoc`.

The tests checking the CSS class names were incorrectly written. Test
inputs used hard coded string instead of the actual value of the various
enum entries. So, they were unable to detect the problem.

BREAKING CHANGES:
- The CSS class name associated to the adhoc subprocess in the DOM have
changed. It was `bpmn-sub-process-ad_hoc` and it is now
`bpmn-sub-process-adhoc`. CSS rules and selectors must be updated
accordingly.
- The value of `ShapeBpmnSubProcessKind.AD_HOC` changed. As it is not
intended to be used directly, no one should be impacted.