test(e2e): restore report generation and fix location of received images
The report was no longer generated and the development server did not stop at the end of the test execution.
This was due to our custom function `toMatchImageSnapshot` becoming asynchronous in commit bffd25b6.
The original function is synchronous, so the custom function must be too.

In addition to this fix, the logic of the code managing image attachments has been improved, mainly to better separate responsibilities.
Storing the expected image copy and calculating the image location are specific to j-i-s, so there's no need to do this. to j-i-s, and are therefore performed first.
The function for attaching images to the report only uses the previously calculated image location and is unaware that j-i-s is being used.

For tests that redefined the location of output images (usually in a subfolder), the path to the received images was not updated accordingly. It was stored in the same location for each test, so only one file was kept and not stored with the diff and the expected image.