refactor: reduce code duplicated from mxGraph to override shape canvas
This reduces the maintenance (in particular if we want to migrate to maxGraph) and reduces the size of the bundle a bit.
The modified code is fully covered by integration tests.
This change will also increase the code coverage as some parts of the duplicated mxGraph code includes paths not used by
`bpmn-visualization`.