refactor!: rename values of multi instance in `ShapeBpmnMarkerKind`
The previous values included spaces which is not consistent with values in other enums.
Use a kebab case and shorten the values.

Refactor also the way markers are painted in `BaseActivityShape`: remove the use of the switch
in favor of the use of a Map.
This makes the implementation consistent with `EventShape`.

BREAKING CHANGES: some values in `ShapeBpmnMarkerKind` have changed.
Values are not supposed to be used directly, so this change should not have any impact.
For applications using such values, replace the value by the of the enum, in particular for:
  - MULTI_INSTANCE_PARALLEL
  - MULTI_INSTANCE_SEQUENTIAL