chore(deps): switch from lodash-es to es-toolkit (#3142)
> es-toolkit is a state-of-the-art, high-performance JavaScript utility
library with a small bundle size and strong type annotations.

Advantages for `bpmn-visualization`:
  - receive more updates
  - optimize for ESM
  - better tree-shaking
- provide types out of the box (for lodash-es, we needed to use a
package provided by DefinitevlyTyped)
- provide CommonJS files as well, so we are no longer forced to use a
trick as we did for lodash-es in the jest configuration of the
integration test (which doesn't provide CommonJS file)

The performance tests were run on Ubuntu 22. 
They include results for master a5886d18 and for switching to
es-toolkit.