test: check text annotation defined in the collaboration node (#3191)
## Overview

We currently only support text annotation defined in process or
sub-process.
Add a visual test to demonstrate that, for now, a text annotation (and
its related association) is not rendered when it is defined in the
"collaboration" node in the XML diagram.
In the BPMN diagram used in the test, there are some text annotations
defined in the "collaboration" that are linked to elements of a process.
So they are not displayed. Such modeling comes from the demo `bpmn-js`
and demonstrates that the limitation doesn't only apply to text
annotation related to pools.