chore(deps-dev): bump typescript from 5.4.5 to 5.8.2 (#3299)
Update configuration and code to ensure compatibility with the new
TypeScript version.

- Use a recent version of `@types/node` that matches the Node.js version
used for building the project.
The previous version had type issues related to `AbortSignal`, similar
to the upgrade to TypeScript 4.9 (see 6356b3e).
- Add an attribute to retain the triple-slash reference to `mxGraph`
types in the type definitions included in the npm package.
As of TypeScript 5.5, this is required; otherwise, the reference is
omitted.


## Notes

This work started in #3124 in August 2024. At that time, we haven't
found in the TypeScript documentation the behavior change about triple
slash references.