fix!: ShapeUtil does not include BPMN artifacts in flow node kinds
Regarding the BPMN semantic, artifacts are not executable flow nodes so they are no longer returned in the result of the
`ShapeUtil.flowNodeKinds` method.

BREAKING CHANGES: artifacts are no longer considered as flow nodes. Code using the `ShapeUtil.flowNodeKinds` method may
need to be updated to concat the array returned by the `ShapeUtil.artifactKinds` method to restore the previous content.