fix!: ShapeUtil does not include BPMN artifacts in flow node kinds  (#3398)
Regarding the BPMN semantic, artifacts are not executable flow nodes so
they are no longer returned in the result of the
`ShapeUtil.flowNodeKinds` method.

BREAKING CHANGES: artifacts are no longer considered as flow nodes.
Code using the `ShapeUtil.flowNodeKinds` method may need to be updated
to concatenate the array returned by the `ShapeUtil.artifactKinds`
method and restore the previous behavior.